% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyfun.R
\name{dummyfun}
\alias{dummyfun}
\title{Dummy function}
\usage{
dummyfun(x)
}
\arguments{
\item{x}{a dummy numeric argument}
}
\value{
Value of `x` increased by 2.
}
\description{
This is a dummy function.
}
\examples{
dummyfun(2)
}
