% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummywidget.R
\name{dummywidget}
\alias{dummywidget}
\title{Joining Animation}
\usage{
dummywidget(col = "red", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{col}{Color of rectangle.}
}
\description{
Joining Animation
}
\examples{
dummywidget(col = "red")
dummywidget(col = "blue")
}
\author{
Charco Hui
}
